Move ThemeManager to proper package
